Cleanup of App.js
