+++ client/.env
@@ -1,2 +1,2 @@
 REACT_APP_API_URL=http://localhost:5001
-REACT_APP_CONTRACT_ADDRESS=0x340E0640e7a893Af39500b5B82b55561582Adb72
+REACT_APP_CONTRACT_ADDRESS=0x165A5BDF5d5B6f6D70b01cB65Ae4b9827cb0B73E
